Renamed materialcomponent to material 2
