+++ app/build.gradle
@@ -63,7 +63,7 @@ dependencies {
     implementation projects.studies.crane
     implementation projects.studies.shrine
     implementation projects.studies.reply
-    implementation projects.studies.materialcomponent
+    implementation projects.studies.material2
     implementation projects.components
     implementation projects.core
     implementation projects.resource
