+++ settings.gradle
@@ -21,7 +21,7 @@ include ':studies:base'
 include ':studies:crane'
 include ':studies:shrine'
 include ':studies:reply'
-include ':studies:materialcomponent'
+include ':studies:material2'
 include ':core'
 include ':components'
 include ':resource'
