+++ studies/material2/proguard-rules.pro
