+++ studies/material2/src/main/AndroidManifest.xml
