+++ studies/material2/src/main/java/com/numero/material_gallery/studies/material2/MaterialComponentFragment.kt
@@ -1,4 +1,4 @@
-package com.numero.material_gallery.studies.materialcomponent
+package com.numero.material_gallery.studies.material2
 
 import android.annotation.SuppressLint
 import android.graphics.Color
@@ -14,7 +14,7 @@ import com.google.android.material.transition.platform.Hold
 import com.google.android.material.transition.platform.MaterialContainerTransform
 import com.numero.material_gallery.core.applyFloatingActionButtonEdgeTreatment
 import com.numero.material_gallery.core.delegate.viewBinding
-import com.numero.material_gallery.studies.materialcomponent.databinding.FragmentMaterialComponentBinding
+import com.numero.material_gallery.studies.material2.databinding.FragmentMaterialComponentBinding
 import dev.chrisbanes.insetter.applyInsetter
 
 class MaterialComponentFragment : Fragment(R.layout.fragment_material_component) {
@@ -66,7 +66,7 @@ class MaterialComponentFragment : Fragment(R.layout.fragment_material_component)
     }
 
     private fun showDialog() {
-        val wrappedContext = ContextThemeWrapper(requireContext(), R.style.Theme_MaterialComponent_DayNight)
+        val wrappedContext = ContextThemeWrapper(requireContext(), R.style.Theme_Material2_DayNight)
         MaterialAlertDialogBuilder(wrappedContext)
             .setTitle("Title")
             .setMessage("Message")
