+++ studies/material2/src/main/res/drawable/ic_material_component_launcher_foreground.png
