+++ studies/material2/src/main/res/menu/menu_mc_bottom_navigation.xml
