+++ studies/material2/src/main/res/mipmap-anydpi-v26/ic_material_component_launcher.xml
