+++ studies/material2/src/main/res/mipmap-hdpi/ic_material_component_launcher.png
