+++ studies/material2/src/main/res/mipmap-xhdpi/ic_material_component_launcher.png
