+++ studies/material2/src/main/res/mipmap-xxhdpi/ic_material_component_launcher.png
