+++ studies/material2/src/main/res/mipmap-xxxhdpi/ic_material_component_launcher.png
