+++ studies/material2/src/main/res/values/array.xml
