+++ studies/material2/src/main/res/values/colors.xml
