+++ studies/material2/src/main/res/values/strings.xml
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <string name="label_material_component">Material Design 2</string>
+
+    <string name="material_component_transition_name">material_2</string>
+</resources>
